// src/app/models/patient.model.ts
export class Patient {
    id: number;
    address: String;
    patientContact: string;
    patientName: string;
    location: String;
    age: string;
    relativeName: string;
    alternateContact1: string;
    alternateContact2: string;
    aadhar: string;
    createdAt: Date;
    addedBy: string;
}